// frontend/src/api/models/groupingResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { FunctionDefinitionMeta } from './functionDefinitionMeta';
import type { GroupingResponseOutputType } from './groupingResponseOutputType';
import type { GroupingStatus } from './groupingStatus';

export interface GroupingResponse {
  assignmentCount: number;
  createdAt: string;
  functionDefinition: FunctionDefinitionMeta;
  groupingId: string;
  groupingMessage: string | null;
  groupingStatus: GroupingStatus;
  jobId: string | null;
  outputType: GroupingResponseOutputType;
  updatedAt: string;
  workflowId: string;
}